requirements.txt
matplotlib==3.9.3
albumentations==1.0.3
torchmetrics==0.10.3
numpy==1.26.4
tqdm==4.67.1
opencv-python